refactor(runtime): extract image_analyze tool module
tool_image_analyze + its three format/dimension helpers
(detect_image_format, extract_image_dimensions, extract_jpeg_dimensions)
and format_file_size move into tool_runner/image.rs. detect_image_format
is promoted to pub(super) because the test block at the bottom of
tool_runner/mod.rs already exercises it directly; the others stay
module-private since they have no out-of-module caller.

The stale "// Location tool" section header that the system extraction
left behind in mod.rs is removed in the same hunk — both Location
tools (location_get and system_time) already live in tool_runner/system,
so the divider was meaningless.

mod.rs: 11,058 -> 10,889 LOC (-169). cargo check + cargo clippy --lib
both clean.

Refs #3710
